Moved Cameras
- Moved camera-related functionality to discrete file
- Fixed issue with system database where new copy was being created due to absolute filepath
